requirements.txt
beautifulsoup4==4.12.3
Bio==1.7.1
biopython==1.83
bioservices==1.11.2
cobra==0.29.1
dash==2.18.2
dash_ag_grid==31.2.0
dash_bootstrap_components==1.6.0
dash_table==5.0.0
dash_uploader==0.6.1
fuzzywuzzy==0.18.0
ipython==8.12.3
langchain==0.3.7
lxml==5.3.0
matplotlib==3.7.5
neo4j==5.26.0
numpy==1.24.4
pandas==2.0.3
py2neo==2021.2.4
python-decouple==3.8
pyvis==0.3.2
streamlit==1.24.0
tomli==2.1.0
